fold hiding logic into central place
