+++ services/ui-src/src/components/fields/RadioField.tsx
@@ -8,6 +8,7 @@ import { ChoiceList as CmsdsChoiceList } from "@cmsgov/design-system";
 import { Page } from "components/report/Page";
 import { ChoiceProps } from "@cmsgov/design-system/dist/react-components/types/ChoiceList/ChoiceList";
 import { requiredResponse } from "constants";
+import { useElementIsHidden } from "utils/state/hooks/useElementIsHidden";
 
 export const formatChoices = (
   parentKey: string,
@@ -60,7 +61,7 @@ export const RadioField = (props: PageElementProps) => {
   }, []);
 
   const [displayValue, setDisplayValue] = useState<ChoiceProps[]>([]);
-  const [hideElement, setHideElement] = useState<boolean>(false);
+  const hideElement = useElementIsHidden(radio.hideCondition);
 
   // Need to listen to prop updates from the parent for events like a measure clear
   useEffect(() => {
@@ -68,23 +69,6 @@ export const RadioField = (props: PageElementProps) => {
       formatChoices(`${props.formkey}`, radio.value, radio.answer) ?? []
     );
   }, [radio.answer]);
-
-  useEffect(() => {
-    const formValues = form.getValues() as any;
-    if (formValues && Object.keys(formValues).length === 0) {
-      return;
-    }
-    if (radio?.hideCondition) {
-      const controlElement = formValues?.elements?.find((element: any) => {
-        return element?.id === radio.hideCondition?.controllerElementId;
-      });
-      if (controlElement?.answer === radio.hideCondition.answer) {
-        setHideElement(true);
-      } else {
-        setHideElement(false);
-      }
-    }
-  }, [form.getValues()]);
 
   // OnChange handles setting the visual of the radio on click, outside the normal blur
   const onChangeHandler = async (
