+++ services/ui-src/src/types/report.ts
@@ -168,6 +168,11 @@ export type PageElement =
   | StatusAlertTemplate
   | DividerTemplate;
 
+export type HideCondition = {
+  controllerElementId: string;
+  answer: string;
+};
+
 export type HeaderTemplate = {
   type: ElementType.Header;
   id: string;
@@ -209,10 +214,7 @@ export type TextboxTemplate = {
   helperText?: string;
   answer?: string;
   required?: boolean;
-  hideCondition?: {
-    controllerElementId: string;
-    answer: string;
-  };
+  hideCondition?: HideCondition;
 };
 
 export type TextAreaBoxTemplate = {
@@ -221,10 +223,7 @@ export type TextAreaBoxTemplate = {
   label: string;
   helperText?: string;
   answer?: string;
-  hideCondition?: {
-    controllerElementId: string;
-    answer: string;
-  };
+  hideCondition?: HideCondition;
 };
 
 export type DateTemplate = {
@@ -282,10 +281,7 @@ export type RadioTemplate = {
   helperText?: string;
   answer?: string;
   required?: boolean;
-  hideCondition?: {
-    controllerElementId: string;
-    answer: string;
-  };
+  hideCondition?: HideCondition;
 };
 
 export type ReportingRadioTemplate = {
