+++ services/ui-src/src/utils/state/hooks/useElementIsHidden.ts
@@ -1,0 +1,25 @@
+import { useEffect, useState } from "react";
+import { useFormContext } from "react-hook-form";
+import { HideCondition } from "types";
+import { elementIsHidden } from "../reportLogic/completeness";
+
+export const useElementIsHidden = (hideCondition?: HideCondition) => {
+  // TODO: Existing bug: Page reload doesn't get set correctly, possible lifecycle issue with form default
+  const form = useFormContext();
+  const [hideElement, setHideElement] = useState<boolean>(false);
+
+  if (!hideCondition) return hideElement;
+
+  useEffect(() => {
+    const formValues = form.getValues() as any;
+    if (formValues && Object.keys(formValues).length === 0) {
+      return;
+    }
+    const hidden =
+      formValues?.elements &&
+      elementIsHidden(hideCondition, formValues.elements);
+    setHideElement(hidden);
+  }, [form.getValues()]);
+
+  return hideElement;
+};
